test_deploy/requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.7.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
